Small refactor of game board
