+++ EngineYardTests/ObsolescenceTests.swift
@@ -10,8 +10,19 @@ import XCTest
 
 @testable import EngineYard
 
+// NEEDS REFACTORING
+
 class ObsolescenceTests: BaseTests {
 
+    override func setUp() {
+        super.setUp()
+    }
+
+    override func tearDown() {
+        super.tearDown()
+    }
+
+    /**
     var gameObj: Game!
     var gameBoard: GameBoard = GameBoard.init()
     var trains: [Train] = [Train]()
@@ -274,6 +285,6 @@ class ObsolescenceTests: BaseTests {
         
         XCTAssert(firstRusted.engineColor == .green)
         XCTAssert(firstRusted.generation == .first)
-    }
+    }**/
     
 }
